commit_74963 ------------------------------------------------------------------------ r74963 | vboxsync | 2018-10-22 16:39:34 +0800 (Mon, 22 Oct 2018) | 10 lines
Additions/linux/vboxvideo: really disable the resize capability on master drop.
bugref:9253: Enabling 2nd virtual screen results in an assert.
The previous change for this bug, by preventing disabling and re-enabling of
acceleration in the device, accidentally prevented propagation of disabling
the resize capability which should happen at the same time, when ownership
of master in the driver changes.  This change explicitly propagates it to the
device.  It fixes a problem whereby the size of the boot splash screen on
shutdown was wrongly remembered as the last size for the virtual machine.  As
the boot splash screen cannot handle dynamic resizing we should ignore any
size it may set for that purpose.
------------------------------------------------------------------------