commit_75323 ------------------------------------------------------------------------ r75323 | vboxsync | 2018-11-08 23:29:56 +0800 (Thu, 08 Nov 2018) | 8 lines
Additions/linux: disable 3D for guests using Wayland.
ticketref:18116: Linux guests with X.Org Server 1.20 and 3D enabled fail...
Our 3D pass-through on Linux guests was designed for an X11 world where GLX
was the only way of accessing OpenGL.  It does not work with Wayland clients
or compositors, and can actively cause problems for EGL with OpenGL.  Case in
point, guests with X.Org Server 1.20 installed do not boot to desktop with
3D enabled.  Rather than working around that for the few remaining GLX
applications on the desktop, just disable 3D when Wayland is in use.
------------------------------------------------------------------------