commit_75399 ------------------------------------------------------------------------ r75399 | vboxsync | 2018-11-12 23:05:43 +0800 (Mon, 12 Nov 2018) | 8 lines
Additions/linux/drm: stop automatically re-indenting kernel module sources.
bugref:8524: Additions/linux: play nicely with distribution-installed Additions
I decided that the minimal benefits of making the vboxvideo sources in the
installer package look more like the kernel ones were not worth additional
complexity in the build process and potential confusion for contributors.
This change removes it, but leaves the sed script in place in case it is
helpful to anyone wishing to compare our sources with the kernel ones, or as
inspiration to anyone needing to do similar things.
------------------------------------------------------------------------