commit_75650 ------------------------------------------------------------------------ r75650 | vboxsync | 2018-11-22 02:30:58 +0800 (Thu, 22 Nov 2018) | 1 line
os2/VBoxSFFile.cpp: Had disabled the HGCM event semaphore change while benchmarking the HLT stuff.  New numbers, gain is 4.8 instead of the 6.4, but that's the joy of benchmarking.  On the plus side, the total gain is now 30 and 38 against the two base line measurements from before the optimization effort. bugref:9172
------------------------------------------------------------------------