commit_75961 ------------------------------------------------------------------------ r75961 | vboxsync | 2018-12-05 17:34:44 +0800 (Wed, 05 Dec 2018) | 12 lines
Additions/VBoxClient/svga: provide default screen layout information.
bugref:9170: Implement support for multihead display configurations in VMSVGA device.
When a multi-monitor guest is in use, the host and the guest have to agree
about how the pointer input device maps to screen layouts in order for input
to be properly usable.  So if the host does not actively request a particular
screen layout we assume screens laid out horizontally starting at zero, which
matches the host default.  I am not sure if the host code will correctly
inform the host input device code if it requests another layout, but if it
does not this works.  And if the guest ignores the host request, e.g. if the
user asked the guest control panel for a different layout, then the input
mapping will break, but this is possibly a guest bug, as it breaks other
virtualisers too.
------------------------------------------------------------------------