commit_76164 ------------------------------------------------------------------------ r76164 | vboxsync | 2018-12-12 00:14:25 +0800 (Wed, 12 Dec 2018) | 10 lines
FE/Qt: for guests using VMSVGA, do not remember guest-initiated resizes.
bugref:5329: Implement VMWare SVGA compatible video device
Currently we remember the last size set by the guest operating system to the
virtual display hardware before it was shut down and resend it on next start-
up.  This is similar to having the same monitor plugged into a physical
system next time it is started.  However, Linux guests using VMSVGA always
set the size to 800x600 with one screen on shut-down, which is not what we
want to remember.  So for guests with the VMSVGA graphics controller type
(normally only used for Linux guests) we only remember sizes which we
explicitly send ourselves, as of this change, to get around this.
------------------------------------------------------------------------