commit_76169 ------------------------------------------------------------------------ r76169 | vboxsync | 2018-12-12 03:19:41 +0800 (Wed, 12 Dec 2018) | 7 lines
FE/Qt: save size and visibility hints to extra data before not after sending.
bugref:5329: Implement VMWare SVGA compatible video device
Since the previous change we save size and visibility hints sent to the guest
in extra data.  However, for non-VMSVGA guests we also save sizes and screen
visibility set by the guest.  This change makes sure that we always save the
ones we send before the ones set by the guest, so that if the guest chooses
to adjust the mode we request we will save the guest choice, not our request.
------------------------------------------------------------------------