commit_76836 ------------------------------------------------------------------------ r76836 | vboxsync | 2019-01-16 20:25:17 +0800 (Wed, 16 Jan 2019) | 2 lines
VMM/IEM: Nested VMX: bugref:9180 Initialize all read-only VMCS data fields to 0 on VM-entry because we don't set every one of these fields on every VM-exit but some of them are not undefined (like Exit-qualification and VM-exit interruption info's valid bit). If we don't initialize these on VM-entry (or on every VM-exit), we risk confusing the nested hypervisor.

------------------------------------------------------------------------