commit_76912 ------------------------------------------------------------------------ r76912 | vboxsync | 2019-01-21 19:21:43 +0800 (Mon, 21 Jan 2019) | 8 lines
Additions/linux: report if the current kernel headers are not installed.
bugref:9364: GAs fail to build and fail to report the failure on a Debian 9.6.
This fixes a regression from previous changes to Additions kernel module
rebuilding.  When the code was generalised to rebuild kernel modules for all
newly installed kernels, not just the current one, the warning when the
headers were not installed was lost, as we do not want to warn when headers
for a random newly installed kernel are missing.  Add an extra check for the
current one to fix this.
------------------------------------------------------------------------