commit_77049 ------------------------------------------------------------------------ r77049 | vboxsync | 2019-01-31 01:06:07 +0800 (Thu, 31 Jan 2019) | 7 lines
Additions/linux/installer: mute chcon error output.
ticketref:18221:Getting /cant apply partial context to unlabeled file/ in Version 6.0.0
We do various chcon invocations in the Additions installer to make SELinux
happy on old Fedora/EL distributions which are mostly forgotten, but on which
people may still try to install.  The invocations are not even correct on
recent systems, so mute the error output.  Do not change them so as not to
break the actual target systems.
------------------------------------------------------------------------