commit_77111 ------------------------------------------------------------------------ r77111 | vboxsync | 2019-02-01 19:01:12 +0800 (Fri, 01 Feb 2019) | 1 line
RTErrConvertFromErrno(): Convert the uNativeCode parameter from unsigned to int. errno is specified to be of type int and while it should only ever contain positive values RTErrConvertFromErrno() is also used inside the various kernel drivers where errno does not exist and the return code might be negative as well. The example here is ERESTART (-1) on FreeBSD causing a compile error with clang and this fix is cleaner IMO instead of casting ERSTART to unsigned. [build fix]
------------------------------------------------------------------------