commit_77130 ------------------------------------------------------------------------ r77130 | vboxsync | 2019-02-02 00:55:20 +0800 (Sat, 02 Feb 2019) | 7 lines
pdmifs/DevVGA/DisplayImpl: add screen id to pfnVBVAReportCursorPosition.
bugref:9376: Complete hardware cursor implementation in VMSVGA.
vboxvideo in Linux passes cursor position information relative to a particular
screen, but we loose this information when we pass it on to the front-end.
Since no front-end has used the information yet this was not noticed.  This
change fixes that and converts the co-ordinates to frame-buffer global ones
as early as possible.
------------------------------------------------------------------------