commit_77193 ------------------------------------------------------------------------ r77193 | vboxsync | 2019-02-07 18:06:19 +0800 (Thu, 07 Feb 2019) | 15 lines
Devices/DevVGA-SVGA: make FIFO polling thread sleep when idle.
bugref:9376: Complete hardware cursor implementation in VMSVGA.
This change adds an access handler for VMSVGA FIFO accesses (the first page
only, which starts with lots of register-type locations before the actual
FIFO begins) which is enabled when we have not seen any activity on the FIFO
for one second or changes to the cursor location registers.  This extends
the old polling on the FIFO with increasing intervals by switching to the
handler when the interval increase reaches a maximum.  After a single access
intercept the handler is disabled again and the polling interval reset to
minimum.  The code contains debug assertions intended to test that we really
do switch correctly in both ways at the right time.
The change also fixes two FIFO access tests to also check the cursor
registers.  Testing was done with DEBUG_FIFO_ACCESS enabled as well, which
also creates a FIFO access handler, to ensure that the changes to not break
that mode.
------------------------------------------------------------------------