commit_77196 ------------------------------------------------------------------------ r77196 | vboxsync | 2019-02-07 21:33:50 +0800 (Thu, 07 Feb 2019) | 6 lines
Help text: explain about EFI Secure Boot and signing.
ticketref:18312: Unable to build Virtualbox 6.0 on Fedora 29
When a user tries to use VirtualBox on a Linux system with EFI Secure Boot
enabled, it sometimes fails in a non-obvious way due to unsigned kernel
modules.  As it is rather tricky to detect this properly, mention it in the
help text which is shown when there are problems with the modules.
------------------------------------------------------------------------