commit_77226 ------------------------------------------------------------------------ r77226 | vboxsync | 2019-02-09 01:13:26 +0800 (Sat, 09 Feb 2019) | 6 lines
Devices/DevVGA-SVGA: move an assertion out of the FIFO access handler.
bugref:9376: Complete hardware cursor implementation in VMSVGA.
I put an assertion into the FIFO access handler that it would not trigger more
than once every 500ms.  However, disabling it is not reliable, so that can
still happen.  Instead, I moved the assertion to the code re-enabling the
handler.  That should really not happen more frequently.
------------------------------------------------------------------------