commit_77265 ------------------------------------------------------------------------ r77265 | vboxsync | 2019-02-11 23:47:23 +0800 (Mon, 11 Feb 2019) | 5 lines
VBoxClient/vmsvga-x11: test for uname(2) being -1 on failure, not just non-null.
ticketref:18428: VBoxClient --vm-svga fails with "Failed to get kernel name" on Solaris guests
On Linux, uname returns -1 on failure and zero on success.  On Solaris it
returns -1 and zero or greater.  So testing it to be non-zero is not valid
for detecting failure.  This change fixes that.
------------------------------------------------------------------------