commit_77451 ------------------------------------------------------------------------ r77451 | vboxsync | 2019-02-25 14:32:10 +0800 (Mon, 25 Feb 2019) | 10 lines
Devices/Graphics, pdm, Main: add out-of-range to host-side pointer reporting.
bugref:9376: Complete hardware cursor implementation in VMSVGA
The front-end can now send an out-of-range pointer event to say that the
pointer is outside of all guest windows.  This can be useful when the cursor
is being drawn in software, as in this case it makes sense to hide it
altogether.  The event is not yet passed down to the guest.  Out-of-range,
like the already existing invalid data event (-1, -1) is passed as special
pointer co-ordinate values to avoid breaking API interfaces if this is
back-ported.  A future change, not to be back-ported, will change the API to
get rid of the special values and do this cleanly.
------------------------------------------------------------------------