commit_77506 ------------------------------------------------------------------------ r77506 | vboxsync | 2019-03-01 00:25:19 +0800 (Fri, 01 Mar 2019) | 6 lines
Main/HostDnsService: fix a use-after-free running VBoxSVC in debug mode.
bugref:9144: AddressSanitizer: heap-use-after-free in VBoxSVC in HostDnsMonitor::pollGlobalExtraData()
This did not trigger during normal usage, but when debugging and keeping
VBoxSVC around it did.  The change terminates the Host DNS monitor thread
properly when the VirtualBox object is destroyed, which in normal use also
ends the process.
------------------------------------------------------------------------