commit_77544 ------------------------------------------------------------------------ r77544 | vboxsync | 2019-03-04 04:07:01 +0800 (Mon, 04 Mar 2019) | 1 line
Runtime/fuzz: Updates, add a target state recording mechanism to record changes in target behavior caused by mutated inputs. This allows to decide which mutated input gets added to the input corpus and which one gets discarded. Currently this is only able to record the stdout/stderr channels of the fuzzed process but other sources to detect changed behvior will get added in the future
------------------------------------------------------------------------