commit_77595 ------------------------------------------------------------------------ r77595 | vboxsync | 2019-03-07 20:42:31 +0800 (Thu, 07 Mar 2019) | 2 lines
VBoxManage: Re-do help category handling (command/sub-command specific help) for old-style commands to lift the low limit of the bit field approach for everything. This is now much closer to the handling of new-style commands, but far from converting anything old. While at it I eliminated the need to define an entry in the (old-style) USAGECATEGORY enum for new-style commands.

------------------------------------------------------------------------