commit_77653 ------------------------------------------------------------------------ r77653 | vboxsync | 2019-03-11 22:03:37 +0800 (Mon, 11 Mar 2019) | 2 lines
vsheriff: Limit check for hanging testboxes to cHoursBack and testboxes which actually existing currently (which will avoid looking at testboxes which are gone for good years ago or are hanging for a while now), which avoids running into errors which prevent re-running the bad testbox checks at the end of the analysis run. This should finally make it work since the new failure rate is rather high and therefore there are usually too many non-categorized failures to make a decision. Also consider driver load failures as a sign of a bad testbox.

------------------------------------------------------------------------