commit_77667 ------------------------------------------------------------------------ r77667 | vboxsync | 2019-03-13 00:03:53 +0800 (Wed, 13 Mar 2019) | 9 lines
FE/Qt: try to avoid saving size hints for guest-initiated resizes.
bugref:9399: FE/Qt: GUI sends size hint for guest-initiated resize.
When the guest resizes this triggers the GUIs window resize code.  To avoid
sending unneeded size hints we check whether the guest screen size already
matches the new window size, the logic being that in that case it was
probably the guest screen resize which triggered the host window resize, and
if it does we do not send a size hint.  In that case though we should also
not remember the size hint for the next time we start the guest.

------------------------------------------------------------------------