commit_77668 ------------------------------------------------------------------------ r77668 | vboxsync | 2019-03-13 00:09:20 +0800 (Wed, 13 Mar 2019) | 5 lines
FE/Qt: Send pointer out of range events to Main.
bugref:9376: Complete hardware cursor implementation in VMSVGA
When the host pointer leaves the GUI window tell Main that it is out of range,
so that the device can hide the cursor if appropriate and if it is drawing it
in software.
------------------------------------------------------------------------