commit_77965 ------------------------------------------------------------------------ r77965 | vboxsync | 2019-03-30 20:35:49 +0800 (Sat, 30 Mar 2019) | 10 lines
Devices/Graphics: more preparations for drawing the pointer cursor in-device.
bugref:9376: Complete hardware cursor implementation in VMSVGA.
This change 1) moves the host cursor moved callback and the host cursor
capability change callback out of the VBVA part of the graphics device and
into the generic part, where they will presumably be implemented.  It 2) also
stops reporting the host cursor position to the guest via VBVA (this was
always treated as optional by the Additions, and a position of (0, 0) treated
as no host support) and always reports that the host can handle hardware
cursor drawing, which is the case with all current front-ends and will be more
so when we can do drawing in the device.
------------------------------------------------------------------------