commit_78048 ------------------------------------------------------------------------ r78048 | vboxsync | 2019-04-09 09:21:09 +0800 (Tue, 09 Apr 2019) | 1 line
IPRT: Working on new RTPathAbsEx implementation, temporarily called RTPathAbsExEx.  This should fix most of the bugs in the current RTPathAbs/Ex code and do away with the fixed path buffer limitations.  Also introduces a RTPATH_BIG_MAX, given that RTPATH_MAX is just a reasonable and not absolute MAX value.  The new one is more or less absolute and must never be used for stack buffers.  bugref:9172
------------------------------------------------------------------------