commit_78272 ------------------------------------------------------------------------ r78272 | vboxsync | 2019-04-24 22:10:29 +0800 (Wed, 24 Apr 2019) | 6 lines
IPRT/rest: RTCRestBinaryResponse::receiveHttpCallback if we are
collecting data in m_pbData we also need to update m_cbData.  I'm not
sure if we want to maintain m_cbDownloaded here at all as it is not
even accessible anyway (except to subclasses), but for now just do a
minimal fix to make the result usable.

------------------------------------------------------------------------