commit_78446 ------------------------------------------------------------------------ r78446 | vboxsync | 2019-05-08 18:57:05 +0800 (Wed, 08 May 2019) | 10 lines
Installer/Linux: remove last uses of KBUILD_EXTMOD to simplify building in-tree.
Currently we use the Make variable KBUILD_EXTMOD to distinguish between builds
inside the Linux kernel kbuild system and outside of it.  However, if our
kernel modules are build inside a kernel tree as part of a kernel build this
does not work, as they are no longer external modules.  So replace those
checks with checks against KERNELRELEASE, and remove other uses of the
variable while we are at it.  Said other uses are inside the host module
top-level Make file.  I am not actually aware of any users of this file,
though if we ever build those modules as part of a kernel build it might be
useful.
------------------------------------------------------------------------