commit_78447 ------------------------------------------------------------------------ r78447 | vboxsync | 2019-05-08 18:59:59 +0800 (Wed, 08 May 2019) | 4 lines
Installers/Linux: make it possible to build our kernel modules in-tree.
This change lets our kernel module build process use kernel headers in a
kernel source tree when they are built as part of a kernel build, and is the
last known requirement for building the Additions modules in that way.
------------------------------------------------------------------------