commit_78698 ------------------------------------------------------------------------ r78698 | vboxsync | 2019-05-24 00:44:05 +0800 (Fri, 24 May 2019) | 1 line
SharedFoldersSvc: Adding function that lets the guest set the error code style.  The idea is that the host will do a some error code adjustments when the guest and host OSes differ, as this is much more effient than the guest doing it.  The deman is chiefly for windows guests.  bugref:9172
------------------------------------------------------------------------