commit_78778 ------------------------------------------------------------------------ r78778 | vboxsync | 2019-05-27 18:08:59 +0800 (Mon, 27 May 2019) | 13 lines
FE/Qt: stop checking for spontaneous resize events.
bugref:9454: FE/Qt: Ubuntu 19.04, mouse-drag resize events not always spontaneous()
ticketref:18677: Can-t Resize display in virtualbox guest
We currently check resize events for spontaneous or not, which is Qt-s way of
telling us whether they came from the window system or not, and as an added
precaution against resize loops only pass them to the guest if they are
spontaneous.  The thinking is that un-spontaneous events will have been
triggered by the guest, so no need to notify about them.  However, on
Xwayland with client-side decorations Qt does the window manager-s job
itself on the client side and sends the resize events itself, without
labeling them as spontaneous, and breaking our logic.
So: this change drops the check, and hopefully other checks which are still
in place should be enough.
------------------------------------------------------------------------