commit_78889 ------------------------------------------------------------------------ r78889 | vboxsync | 2019-05-31 17:00:08 +0800 (Fri, 31 May 2019) | 3 lines
VMM/IEM: Nested VMX: bugref:9180 Get the IDT-vectoring info. from IEM during VM-exit rather than recording it in the VMCS directly during event injection and then trying to clear it on VM-exit when its not applicable.
This is so that hardware-assisted VMX execution can pass on the IDT-vectoring info. as part of the VM-exit without the IEM VM-exit code overwriting it.

------------------------------------------------------------------------