commit_79073 ------------------------------------------------------------------------ r79073 | vboxsync | 2019-06-11 13:26:37 +0800 (Tue, 11 Jun 2019) | 2 lines
VMM: Nested VMX: bugref:9180 With HM execution of nested-guests, we merge the guest/host mask of the guest and nested-guest. Hence, while constructing the read-shadow, we must use the merged mask. Pass the merged mask as a parameter to CPUMGetGuestVmxMasked[Cr0|Cr4] as otherwise it will continue to use the original mask from the nested-guest VMCS instead.

------------------------------------------------------------------------