commit_79383 ------------------------------------------------------------------------ r79383 | vboxsync | 2019-06-27 17:53:17 +0800 (Thu, 27 Jun 2019) | 2 lines
VMM/HMVMXR0: Nested VMX: bugref:9180 Two more cases (Apic-access and task switch VM-exit) of zero'ing out the wrong struct (copy/paste bug). Fix passing the VM-exit reason to Mov DRx Vm-exit handling, remember 0 is a valid reason, so we can't easily verify if the reason is valid or otherwise if we fail to pass it.

------------------------------------------------------------------------