commit_79403 ------------------------------------------------------------------------ r79403 | vboxsync | 2019-06-28 17:46:46 +0800 (Fri, 28 Jun 2019) | 10 lines
FE/Qt: UIMachineViewNormal: base adjust guest screen size on last size hint.
bugref:9482: FE/Qt and Additions/WINNT: move away from *adjust size...*
At regular intervals the GUI performs checks to see if it should send the
guest machine a new preferred size.  Up until now this has been done if the
current size of the guest screen is too large for the host screen it is
located on.  However this triggered unwanted resizes as the check was also
performed when the guest enabled dynamic resizing, but before it was able
to set a size itself.  This change modifies the behaviour, so that the GUI
now checks whether the last set preferred size for the guest screen will fit
on the host screen.
------------------------------------------------------------------------