commit_79427 ------------------------------------------------------------------------ r79427 | vboxsync | 2019-06-29 08:06:56 +0800 (Sat, 29 Jun 2019) | 9 lines
IPRT/rest: Fix inverted logic to specify kDoCall_RequireBody flag.  We
need it precisely when the body parameter is *not* required at the
REST level - we need to sign the body even if it's empty.

Regen the C++ API sources, where InstanceAction is now the only
operation affected by the (cumulative) mustache changes.

bugref:9436 bugref:9167

------------------------------------------------------------------------