commit_79489 ------------------------------------------------------------------------ r79489 | vboxsync | 2019-07-03 16:02:20 +0800 (Wed, 03 Jul 2019) | 2 lines
VMM/HMVMXR0: Nested VMX: bugref:9180 Don't add/remove the TSC_AUX MSR during VM-entry because we might transition to executing the nested-guest after the VM-exit. So add during VM-entry and remove on the VM-exit to make sure the MSR bitmap permissions are what they should be in case we transition to executing a nested-guest.

------------------------------------------------------------------------