commit_79611 ------------------------------------------------------------------------ r79611 | vboxsync | 2019-07-09 07:33:59 +0800 (Tue, 09 Jul 2019) | 1 line
VBoxManage,manual: Added the dhcpserver subcommand findlease for testing IDHCPServer::FindLeaseByMAC().  Wrote a man page for the dhcpserver command and converted the help to the new style.  Also took a shot at simplifying the option handling of the existing dhcpserver subcommands while retaining a somewhat more flexible compatibility with the old syntax (not advertised) - not really tested.  bugref:9288 bugref:9151
------------------------------------------------------------------------