commit_79640 ------------------------------------------------------------------------ r79640 | vboxsync | 2019-07-09 19:22:28 +0800 (Tue, 09 Jul 2019) | 7 lines
FE/Qt: X11 hosts: ask the X server if we have the focus before grabbing.
bugref:9506: FE/Qt: SUSE report GUI problem with Qt 5.13.
Changes in Qt 5.13 which delay reporting focus out events on X11 upset our
keyboard grabbing code, which waits for a short delay after receiving focus
to grab the keyboard and checks if we still have it.  The delay in reporting
focus out defeats this check, so with this change, on Qt 5.13 and later we ask
the X server directly, not via Qt.
------------------------------------------------------------------------