commit_79648 ------------------------------------------------------------------------ r79648 | vboxsync | 2019-07-10 12:54:24 +0800 (Wed, 10 Jul 2019) | 2 lines
VMM/HMVMXR0: Nested VMX: bugref:9180 Fix TSC offsetting when using nested-guests. We should only remove the RDTSC exiting control when -both- the guest and nested-guest allow it. Fixed statistics on TSC offset vs intercept to be more useful (not just when changing controls which is optimized to avoid vmwrites when unneeded and on CPU migrations etc).  Use the new macros added in r131955.

------------------------------------------------------------------------