commit_79966 ------------------------------------------------------------------------ r79966 | vboxsync | 2019-07-25 04:38:10 +0800 (Thu, 25 Jul 2019) | 1 line
Storage: Added a desired format parameter to VDGetFormat() so Main can pass along the device type.  Bumped the RAW backend down after the CUE and VISO to prevent (impossible) mixups of tiny files.  Extended rawProbe() to look for a valid ISO-9660/UDF descriptor sequence on the image if the caller doesn't say it should be a floppy or hdd, only if that fail go by the extension.  Note! the pfnProbe callback should probably get a score return parameter to indicate how confient the backend is about the probe result, as this would prevent the RAW backend from accidentally grabbing images which belongs to a plug-in.  bugref:9151
------------------------------------------------------------------------