commit_80510 ------------------------------------------------------------------------ r80510 | vboxsync | 2019-08-30 17:56:16 +0800 (Fri, 30 Aug 2019) | 2 lines
VMM/IEM: Added an assertion in IEMExecLots that the TRPM event being checked is indeed an interrupt (since we also check if interrupts are enabled before injecting them). Otherwise, we risk not delivering an exception (that is in TRPM) if interrupts are disabled and can lead to invalid execution sequence.

------------------------------------------------------------------------