commit_81374 ------------------------------------------------------------------------ r81374 | vboxsync | 2019-10-19 05:47:55 +0800 (Sat, 19 Oct 2019) | 1 line
iprt/cdefs.h: doxygen markup interfers with @def __X86__, __AMD64__ and __PRETTY_FUNCTION__, as it takes the double underscores to indicate <strong>X86</strong>.  Newer doxygen seems to allow \__X86__ and such, but the one we have on the build box doesn't grok that.  So, deal with it differently to avoid build failures.
------------------------------------------------------------------------