commit_81501 ------------------------------------------------------------------------ r81501 | vboxsync | 2019-10-24 09:37:27 +0800 (Thu, 24 Oct 2019) | 1 line
VMM/IEMAllCImplStrInstr.cpp.h: Fixed a few incorrect uses of OP_SIZE in aligment mask calculations, code thought it was a byte count when it's a bit count.  Also added a optimization for 'rep stos' on unassigned pages as some buggy EFI firmware ended up doing that for hundred of MBs, which takes forever when doing it byte-by-byte of course.
------------------------------------------------------------------------