commit_81581 ------------------------------------------------------------------------ r81581 | vboxsync | 2019-10-30 17:45:55 +0800 (Wed, 30 Oct 2019) | 1 line
Main,FE/VBoxManage: Add an option to present the SMBIOS System UUID (SMBIOS spec chapter 7.2.1) in little endian format to the guest like it is supposed to be done. To not break existing behavior and causing certain software to require re-activation existing VMs will keep the current (wrong) beahvior. New VMs will be switched to the correct behavior. In addition VBoxManage got a new parameter to set the behavior evene for existing VMs
------------------------------------------------------------------------