commit_81751 ------------------------------------------------------------------------ r81751 | vboxsync | 2019-11-08 23:01:34 +0800 (Fri, 08 Nov 2019) | 1 line
DevLsiLogicSCSI: Replaced the three dynamically allocated queues with static arrays in the device instance data.  The queue size can be configured by VBoxInternal trickery, so we'll typically always be using the defaults of 256+1, so no real need to use the hyper heap here (we want to avoid it due to cross ring pointer resolving fun).  bugref:9218
------------------------------------------------------------------------