commit_82252 ------------------------------------------------------------------------ r82252 | vboxsync | 2019-11-28 05:31:53 +0800 (Thu, 28 Nov 2019) | 1 line
vmm/pdmaudioifs.h: Style, docs and other nits. First, it's always _FLAGS_ never _FLAG_. Second, enums generally should start with _INVALID = 0 to ensure we don't mistake zero-initialized memory for valid data. Struct member names shall be indented on a tab (+4) boundrary.  PDM is part of the VMM, so it follows the VMM coding guidelines strictly. Skip the 'Structure for keeping a ... around' fluff, the first sentence of a structure (or anything else for that matter) documentation shall be brief and to the point.  It is automatically turned into a @brief.  Furthermore, additional text should be a separate paragraph as it provides details the reader doesn't necessarily need to read.  bugref:9218
------------------------------------------------------------------------