commit_82254 ------------------------------------------------------------------------ r82254 | vboxsync | 2019-11-28 06:09:17 +0800 (Thu, 28 Nov 2019) | 1 line
vmm/pdmaudioifs.h: The prefix 'cf' reads 'count-of-flags', so if you want 'count-of-frames' in a non-confusing way you have to spell it out.  The prefix 'cx' means 'count-of-x-coordinates', i.e. width in pixels, and is not in any way suited to combine 'cb' and 'cFrames' in this context (PDM). bugref:9218
------------------------------------------------------------------------