commit_82265 ------------------------------------------------------------------------ r82265 | vboxsync | 2019-11-28 19:22:57 +0800 (Thu, 28 Nov 2019) | 2 lines
ACPI/vbox-cpuhotplug.dsl: Workaround for recent iasl failing to compile the CPU hotplug SSDT because of strings used as _UID for the processor (the ACPI spec seems to avoid defining the type, but iasl sometime after 20190703 insists it must be an integer). Simplify the macro by using the first parameter "id" also for the _UID of the processor.

------------------------------------------------------------------------