commit_82308 ------------------------------------------------------------------------ r82308 | vboxsync | 2019-12-01 09:20:59 +0800 (Sun, 01 Dec 2019) | 1 line
DevHDA: Fixed bug in hdaCodecLoadState showcasing why switch aren't the best way of handling saved state versions.  If the switch has current-version case, it's either already broken or just waiting for it to happen.  (Current version shouldn't be used anywhere but in the registration statement.)  This one broke when the state version was last bumped.  bugref:9218
------------------------------------------------------------------------