commit_82401 ------------------------------------------------------------------------ r82401 | vboxsync | 2019-12-05 07:33:29 +0800 (Thu, 05 Dec 2019) | 1 line
DevHDA: Handle some partial/multiple register accesses (both reads and writes) in ring-0.  In the read case IOM may upon our request translate a byte or word access that exactly matches the desired register into a dword access spanning multiple registers.  We can only safely handle these if we know there won't be any VINF_IOM_R3_MMIO_READ statuses showing up.  Fortunately we can tell that from the pfnRead functions.  In the write case, the access may be a byte write to a 16-bit, 24-bit or 32-bit register.  We just supply the missing bits and do it.  Mind we only support doing this if it's the first byte in the register that's being accessed, but that seems to do the trick almost all of the time.  bugref:9218
------------------------------------------------------------------------