commit_82481 ------------------------------------------------------------------------ r82481 | vboxsync | 2019-12-07 09:05:01 +0800 (Sat, 07 Dec 2019) | 1 line
VBoxClipboardSvc: Moved the saved state version defines from the public service header and into the service implementation where it ough to be.  Also, s/_SSM_/_SAVED_STATE_/ so the things make sense (the saved state manage doesn't have a version, only an $Id$ at the top of the file), and s/_VER_LATEST/_VER_CURRENT to conform with the rest of VBox.  Explained why bit 31 is set in the version number as the code related to this was apparently jettisoned recently.  bugref:9437
------------------------------------------------------------------------