commit_82494 ------------------------------------------------------------------------ r82494 | vboxsync | 2019-12-07 20:01:19 +0800 (Sat, 07 Dec 2019) | 1 line
VBox/HostServices/VBoxClipboardSvc.h: Moved the protocol change log up to the @file section as doxygen would attach it to the next thing following it, like the VBOX_SHCL_MODE_XXX section.  Also edited it as it sounded like a sales pitch for the new protocol and a bashing of the previous stuff, which isn't very helpful nor much of a changelog/history.  What's there now is incomplete, but at least it says so.  bugref:9437
------------------------------------------------------------------------