commit_82525 ------------------------------------------------------------------------ r82525 | vboxsync | 2019-12-10 03:48:38 +0800 (Tue, 10 Dec 2019) | 1 line
SharedClipboardSvc,Vbgl: Looked at the connect and feature exchange stuff. Decided to make the connect one always fail, that way the beta and rc additions should not try make use of any of the new features and we should be free to do whatever we like.  Need to circle back and drop the _61B parameter count variants later.  bugref:9437
------------------------------------------------------------------------