commit_83433 ------------------------------------------------------------------------ r83433 | vboxsync | 2020-03-26 08:01:16 +0800 (Thu, 26 Mar 2020) | 7 lines
OCI: we cannot return an array under progress.  Reshuffle the code so
that readCloudMachineList reads the list from the cloud and then (when
the progress is completed) the list can be read from the
cloudMachineList property.  A bit ugly, but the alternative is to
introduce wrapper/holder objects (like IStringArray) which is a lot of
boilerplate code.

------------------------------------------------------------------------