commit_83451 ------------------------------------------------------------------------ r83451 | vboxsync | 2020-03-27 03:17:38 +0800 (Fri, 27 Mar 2020) | 1 line
TestManager/common.js: The iframe#rXXXXXX parent scroll/jump thing should be fixed now. Need to keep it all hidden till the iframe is fully loaded and only then display it. Optimized the changing of the #rXXXXXX part, though not for firefox which seems to not want to act on location.hash changes.
------------------------------------------------------------------------